Changes to add ability to append extra information to the end of URLs and have the server make it available.  Info is available through the function request.extraPathInfo()
Application.py
modified serverSidePathforRequest to extract the extra path info when configured to permit this.

HTTPRequest.py
modified to add the variable self._extraPathInfo to the class, added function extraPathInfo whcih returns this value, added serverURL() function which returns the full internet URL of the request, without the httP://, and serverURLDir, which returns the full internet URL of the request, but only up to the directory.  This is useful for setting a <BASE HREF+"..."> when using extra path info URLs so that relative links can be used.

Configs/Application.config
Added configuration switch for ExtraPathInfo

Configs/Application.config.default
Added configuration switch for ExtraPathInfo