Added new serverSideInfoForRequest algorithm/method
Set UseNewPathAlgorithm in Application.config to enable this new
algorithm (the old algorithm is still in place and will still
be used by default).

The new algorithm always uses extraPathInfo, though perhaps it should
be crippled if the setting is off (there shouldn't be a significant
performance hit either way).  The logic for the FilesToHide/FilesToServe
settings have been moved directly into filesnamesForBaseNameNew (the new
equivalent of filesnamesForBaseName).