Refactored HTTPAdapter, integrating ideas from HttpAppServer. There's an HTTPServer.py module which contains most of the logic for HTTP serving (taken from HTTPAdapter.py).
HTTPAdapterServer takes the place of HTTPAdapter, and serves HTTP
and connects to the AppServer via the adapter interface.

HTTPEmbeddedServer puts the server in the same process as the AppServer,
and communicates directly with the AppServer.  It can be used in
conjunction with ThreadedAppServer, or potentially it could take its
place.  It doesn't have any of the threading management like
ThreadedAppServer, but it has all the same basic functionality.

There's no way to install the embedded server yet.  Stay tuned!