Created new, multi-port application server.  Should be moved back into ThreadedAppServer once it's better tested.
A handlers for the traditional adapter service is located in
NewThreadedAppServer.py, as well as the monitor service
(totally untested).

HTTPServer implements a HTTP server.  Use:
  ./AppServer http
to enable it.  The adapter interface will remain.

The HTTPAdapter has been refactored so that it shares code with
the built-in server, and moved to Adapters/.  The adapter has an
advantage of not dropping requests when the AppServer restarts
(which the built-in version necessarily will do).