Fixes patch [ 642612 ] FancyTraceback fails if source missing
inspect.getframeinfo must always return a filename.  Previously if the
file did not exist, it was returning None.  The frame source file may
not always exist as in the case of errors in pre-compiled
installed libraries.

Updated exception handler in case there is an error printing
the fancy traceback it will do something useful.