Updated to use "cvs export" by default.  Has the benefit of making it simpler to verify that what is in the released tarball is only what came from CVS, and does not include local working directory changes that may not have been committed, or files that may have been added without being placed in CVS.
Takes a CVS tag as an optional argument that will allow exporting
the specific tag.  This is the method to use to do a full release.
Snapshots will be the default with a datestamp.

If you want the old behavior, using an updated current directory
pass the argument "method=current".