* Application rewrite.  Much of the functionality is moved to URLParser,   or into the request/response objects. * URLParser module.  Uses parser objects that delegate to each other.
The doc strings more or less explain the new architecture.  The ultimate
goal is to make it possible to adapt parts of the system more easily,
for instance to make URLs in some trees act differently.

Also HTTPExceptions support, where some cases (like file-not-found) are
handled with exceptions.  Any non-200 (or non-2xx) response code should
be handled with an exception.