Added assertion checks to detect when a changed object is about to be refreshed from the database (changes would be lost).  See release notes for details.
In order to detect this reliably I tweaked the logic in the generated code
which sets MiddleObject._mk_changed, so that it only gets set when an attribute
actually changes.  Hopefully this doesn't break anyone's logic, but IMO the
old semantics for this flag were pretty much useless, and this needs to be
fixed at some point.