I'm using apache's mod_gzip-1.3.26.1a in front of Webware. When fetching static files that are processed by Webware and mod_gzip, I noticed that the HTTP headers contained: 	content-length: A 	Content-Length: B where A is the original file size, and B is the gzipped size. Needless to say, two such headers are illegal.
This is caused by a bug in mod_gzip: it is supposed to replace
the original Content-length header, but fails to recognize
the lower-case version, even though HTTP headers are case-insensitive.
I've submitted a patch for that.

This patch modifies HTTP headers generated by WebKit so that
the first character is capitalized, which works around the mod_gzip bug.