Bugfix in SessionStore as suggested by Ben Parker on webware-discuss 2006-04-03:
"The SessionStore convenience functions items() and values() include a bug where a session key can be deleted after self.keys() has been called, but before the session has been accessed from the store. This leads to an unhandled KeyError which causes SessionSweeper to fail, allowing sessions to pile up until catastrophic failure due to lack of memory.

When using the SessionMemoryStore, this bug is extraordinarily rare. I only encountered it with a high-traffic site where we use SessionDynamicStore, and every Monday (the site's highest traffic day) the appserver would consume all memory ending up with about 75,000 sessions, and I would find traces of this error in the logs (way back in the logs). After applying the fix a few weeks ago, the app has been fine ever since."

I have added a comment and changed the get() method in a similar way.